condense config files
